07:23 am, we stand at dawn

+++ app/src/main/java/edu/thunderseethe/rapulazer/AudioLib/AudioFeatures.java
@@ -1,8 +1,5 @@
 package edu.thunderseethe.rapulazer.AudioLib;
 
-import android.util.Pair;
-
-import java.util.ArrayList;
 import java.util.Locale;
 
 import be.tarsos.dsp.pitch.PitchDetectionResult;
@@ -36,7 +33,7 @@ public class AudioFeatures {
      * the pitch itself.
      */
     public PitchDetectionResult pitch_detection_result;
-    public int[] freq_counts;
+    public double[] freq_counts;
 
     public String toString() {
         return String.format(Locale.CANADA, "AudioFeature(%b, %b, %b, %f, %f, %f)", is_silence, is_beat, is_percussion, sound_pressure_level, average_envelope, pitch_detection_result.getPitch());
